requirements.txt
PyYAML
awesome-slugify
